refactor: change px to rem
